tests/requirements.txt
# Test dependencies for Symphonic-Joules
pytest>=7.0.0
pytest-cov>=3.0.0
PyYAML>=5.1